+++ client/src/app/dashboard/dashboard.component.html
@@ -1,2 +1,2 @@
-<h1>My beautiful header</h1>
+<h1>Delphi</h1>
 <router-outlet></router-outlet>
