+++ client/src/app/dashboard/dashboard.module.ts
@@ -6,7 +6,7 @@ import { DashboardComponent } from './dashboard.component';
 import { CrawlerComponent } from './crawler/crawler.component';
 import { WebappComponent } from './webapp/webapp.component';
 import { WebapiComponent } from './webapi/webapi.component';
-import { DashboardOverviewComponent } from '../dashboard-overview/dashboard-overview.component';
+import { DashboardOverviewComponent } from './dashboard-overview/dashboard-overview.component';
 
 @NgModule({
   imports: [
